Move os arquivos de model para a camada dominioo
