+++ EFCore.Dominio/Arma.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace EFCore.Dominio
+{
+    public class Arma
+    {
+        public int Id { get; set; }
+        public string Nome { get; set; }
+        public Heroi Heroi { get; set; }
+        public int Heroiid { get; set; }
+    }
+}
