+++ EFCore.Dominio/IdentidadeSecreta.cs
@@ -3,13 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace EFCore.WebApi.Models
+namespace EFCore.Dominio
 {
-    public class Arma
+    public class IdentidadeSecreta
     {
         public int Id { get; set; }
-        public string Nome { get; set; }
+        public string NomeReal { get; set; }
+        public int HeroiId { get; set; }
         public Heroi Heroi { get; set; }
-        public int Heroiid { get; set; }
     }
 }
