--- EFCore.WebApi/Models/HeroiBatalha.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace EFCore.WebApi.Models
-{
-    public class HeroiBatalha
-    {
-        public int HeroiId { get; set; }
-        public Heroi Heroi { get; set; }
-        public int BatalhaId { get; set; }
-        public Batalha Batalha { get; set; }
-    }
-}
